+++ backend/src/main/java/org/scoula/safety_inspection/infra/cors/service/CopyOfRegisterMultiService.java
@@ -96,7 +96,7 @@ public class CopyOfRegisterMultiService {
         String commonOwner = getCommonOwner(firstEntry);
         String changeOwnerCountStr = getChangeOwnerCount(firstEntry);
         Integer changeOwnerCount = Integer.parseInt(changeOwnerCountStr);
-        Integer maximumOfBond = Integer.parseInt(getMaximumOfBond(firstEntry));
+        Long maximumOfBond = Long.parseLong(getMaximumOfBond(firstEntry));
 
 
         CopyOfRegisterDto registerDto = new CopyOfRegisterDto(
